Added content to servlet "Payment".
1. Added a javascript function to validate users input in payment.jspx
page, which dose not allow user to leave any empty input text field
before submit.

2. Changed name of "addToCart" button and name of quantity input field
in "viewAll.jspx" and "searchedBooksDisplay.jspx". Originally, the names
are default values which will deliver wrong values to the servlet.

3. Added content to servlet "Payment.java". Now the servlet is able to
process user's input and create corresponding AddressBean objects and a
POBean object. Furthermore, the servlet will save the data into
database.

4. Hard coded that every 3rd order is labelled with status "DENIED".

5. All other orders are simply labelled with "PROCESSED". In other
words, further implementation is needed to set an order as "ORDERED"
first in servlet "OrderProcessCtrl.java" and then in servlet
"Payment.java" change the status of the order to "PROCESSED" (if its is
not a denied order).