Website Functionalities Nearly Completed.
1. Accomplished functionality of user registration. Servlet "Register"
receives request from register.jspx and stores user input into database.
Also, if provided username already existed in table "Customer", the
registration process would fail.

2. Modified navigation bar to have two versions. It depends on whether
there is a logged in user. This is achieved by adding c:if conditions in
original "navbar.jspx" file.

3. A new "Message page" (message.jspx) is created. UI IS EXTREMELY
DISAPPOINTING!! The purpose of this page is to give user a feedback
after any logging or registration behaviors. The message page consists
of two parts: a message body and a go-back button. The message body
displays the message which is specified in servlets (Login.java and
Register.java) depends on result of corresponding behavior (successful
or failed). The go-back button redirects the page to the previous page
which user was browsing. The uri of the button is retrieved in one of
the servlets by getHeader("referer") method.

4. Added connection between "my account" button with "My Account" page.

5. Accomplished functionality of updating user profile in "My Account"
page. Added interaction with database in "OrderHistory" (SHOULD CHANGE
TO A MORE PROPER NAME) servlet.

6. Discarded Ajax function "deleteBook()" in "Login.js" (ALSO CHOOSE A
PROPER FILE NAME PLZ). Since originally, if Ajax was used as a method to
interact with "OrderProcessCtrl" servlet, a more complicated Ajax method
would be created to update real-time shopping cart info, however we
almost have no time for this kind detail implementation. As a result,
the interaction method changed to submitting a form.

7. Added input validation for almost all input textfields (should be
checked again) in "my account" page and "registration" page, which means
user can not submit a form if any of inputs is empty. Again, the reason
is no time for detail implementation.

8. "Edit Account Info" panel now can display user's default first name,
last name and email address.

9. Added a simple javascript funtion "confirmSignOut()" to prompt user
when they clicked on sign out button on navigation bar.

10. "signedInNavbar" now is useless and can be discarded.